// lib/supabase/types.ts
export type Database = {
// ...
  public: {
    Tables: {
      bed_events: {
        Row: {
          bed: string
          created_at: string
          end_time: string
          id: string
          patient_id: string
          start_time: string
          updated_at: string
        }
        Insert: {
          bed: string
          created_at?: string
          end_time: string
          id: string
          patient_id: string
          start_time: string
          updated_at?: string
        }
        Update: {
          bed?: string
          created_at?: string
          end_time?: string
          id?: string
          patient_id?: string
          start_time?: string
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "bed_events_patient_id_fkey"
            columns: ["patient_id"]
            isOneToOne: false
            referencedRelation: "patients"
            referencedColumns: ["id"]
          },
        ]
      }
      board_beds: {
        Row: {
          bed: string
          created_at: string
          sort_order: number
        }
        Insert: {
          bed: string
          created_at?: string
          sort_order?: number
        }
        Update: {
          bed?: string
          created_at?: string
          sort_order?: number
        }
        Relationships: []
      }
      board_plans: {
        Row: {
          created_at: string
          id: string
          text: string
          time: string
          updated_at: string
        }
        Insert: {
          created_at?: string
          id: string
          text: string
          time: string
          updated_at?: string
        }
        Update: {
          created_at?: string
          id?: string
          text?: string
          time?: string
          updated_at?: string
        }
        Relationships: []
      }
      patient_list_entries: {
        Row: {
          collection: string
          created_at: string
          from_time: string
          id: string
          patient_id: string
          to_time: string
          updated_at: string
        }
        Insert: {
          collection: string
          created_at?: string
          from_time: string
          id: string
          patient_id: string
          to_time: string
          updated_at?: string
        }
        Update: {
          collection?: string
          created_at?: string
          from_time?: string
          id?: string
          patient_id?: string
          to_time?: string
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "patient_list_entries_patient_id_fkey"
            columns: ["patient_id"]
            isOneToOne: false
            referencedRelation: "patients"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}